FIX: Issue #34  Editable variables aren't removed
Previous behaviour: CellEditCallback function updated the cfg, but was only called when the user edited any cell

The Add,Remove and Clear buttons modified the data in the object, but not the cfg. If the user then edited a cell the cfg would be updated, but otherwise the cfg would not be updated.

Solution: listener added to object.data on PostSet, which calls the method in CellEditCallback

New Behaviour: cfg updated
1. when Add, Remove or Clear buttons are clicked
2. when any cell is edited