FIX: sessions not properly ended on 2014a
MATLAB 2014a declares handles as invalid on exit. Using `onCleanup` with an `exit` statement in the function may therefore cause problems with IPC via pipes. Exit was therefore moved outside of runCore.m into start_arcade.bat.

On 2014a with 2.5.0 and 2.5.1, eye data was therefore not received properly. The error does not occur with 2018b.